backend/requirements.txt
ipython
langchain-community
langchain-openai
langgraph
langgraph-checkpoint
python-dotenv
requests
Flask
openai
flask-cors
